fix QuickWhois for gfaqs9
gfaqs9 uses the old 'Public E-Mail Address' and 'IM' fields. There are two
ways we can deal with this. We can either detect the old fields, or we
can just make the request to /gfaqs/. Both have their downsides (either
we display probably outdated information, or we take the risk of just
blindly changing the link and hoping it works). Option 2 sounds good to me.