Cache page type in the document
The thing to watch out for is overlapping pages. Message detail is (correctly)
detected as a message list, but it has its differences and must also be
detected as detail.php. This is accounted for by using an array -- if the page
is a message list, also check if it's detail.php and store both.