Fix truncated HTML entities in signatures
I've tested this with a lot of things and it seems to work.

I've noticed that QuickPost has a problem with HTML characters in signatures
because it's generated with innerHTML - you can easily break things by putting
</textarea> in your sig, and any entities you use will be parsed. This can be
fixed by using the DOM, but for now the signature should be filtered through
specialCharsEncode().