Smaller delay in count
1200 is pretty conservative -- you can type a whole paragraph quickly
and not get any feedback. 100 seems to update after 5 or so letters,
at least at my typing speed. I think this is a nice medium between
constant updating and not getting any response with fast typing.

Of course, the whole concept of a timer is conservative, but while there
wasn't any huge slowness before, it was definitely using a lot more CPU
than it had to be.

Also, move the new functions inside the object. Declaring them outside
it is pretty messy.