Use DOM to create the QuickPost form
Consequently, the reset button broke (completely empties the textarea),
so I just removed it... I don't think anyone will care.

There's some issues with spacing between buttons now, which requires
some text nodes in between them to solve.

Also, the title above the form was removed because it wasn't that useful
and it looked ugly - but I don't really care if it stays or goes.

The QuickPost CSS will be updated later.