Better QuickPost page redirection
Previously, QuickPost always loaded the page from which you posted
the message. This starts to become a problem with topics with multiple
pages.

* Always load the last page of a topic after QuickPosting - this is
  determined using the page jumper
* If the user is on the last page of the topic and it's full, load the
  next (not-yet-existent) page to see the newly posted message

This works best when the user is on the last page of a topic - on the
first page, the number of messages in the topic can't be determined.

Also, I haven't had it jump to the last post using the hash because the
page won't reload if the page stays the same. There should be some way
to work around this, though, because that'd be really convenient...

To more neatly allow for storing things in the document, I added a
doc.gamefox object. GFlib.onPage should be updated to use this.