8for11 update
Takato:
- If you have a board on your Board List that requires you to be at a higher
user level in order to view it, previously there was a bug where the board
name disappeared. This is now fixed, and the name of the board is now
visible again.
- Most users would have only noticed this while suspended/warned/locked or
similar. Long-time users may have also noticed it with boards they
favourited back when you did not need to meet a level requirement to
favourite a board. Also, any ex-moderators who still have mod-only boards
favourited would have noticed this.