Update release scripts
I don't like the idea of putting XPIs in git, and this method seems very
hackish. Now that I have a VPS, more things are possible with release
management if the need ever arises. For now, keeping the McCoy profile
local and logging into the server directly to upload makes the most
sense.