Initial Firefox 4 compatibility
Added support for the new AddonManager API (used for getting the version at
startup).

I moved the update code into the gamefox_update() function, so it can both be
used as a callback function in Firefox 4 and as a normal, synchronous function
in earlier versions.

The version-getting code in gamefox_lib now only checks the gamefox.version
preference. This ensures that the variable is set instantly, with the only
downside being that it might be outdated for a few milliseconds after an
upgrade until gamefox_update() updates the version.

I moved uG from em:creator to em:developer in install.rdf because the Fx4
addon manager was saying "by Abdullah A", which isn't accurate. Right now
em:creator is blank, but maybe it should be "GameFOX developers" or something.

TODO: Find out why the page jumps to bottom (around the quick nav and
QuickPost) on message lists