Fix autoscrolling to bottom of message list in Fx4
setSelectionRange() is the culprit here. I've moved it to an event listener
that fires when QuickPost is focused, so it won't cause Firefox to scroll as
soon as the page loads.

Quoting isn't affected because it detects if the cursor is in the signature
(the default without setSelectionRange) and inserts the quote at the beginning
of the message. The event listener deletes itself after running once, so the
cursor position won't be lost if the user unfocuses and tabs back into
QuickPost.