Move lib version functions to version object
Mostly a cosmetic change for the code. I never liked having the version
functions grouped with all of the other gamefox_lib functions.

Also removed some unused functions in the process (getBaseVersion,
getNightlyAge.

A disadvantage is that gamefox_lib.version is no longer a string. By defining
the toString and indexOf functions, this doesn't matter for our purposes,
except for one line in overlay.js. gamefox_lib.version._version is ugly, but I
don't think it's possible to define getters and setters for objects in
JavaScript, only properties of objects.