Fix options window being modal with instantApply
On Windows, prefwindows are modal dialogs, but on Linux, they aren't.
Unfortunately, our function for opening the options window didn't consider
this, and made it modal on all platforms. This commit fixes that by checking
the instantApply pref. It now behaves identically to how the Firefox addons
manager opens preference windows.