New highlighting interface
This is a two pane UI instead of the list of groupboxes we had before.

This is a much cleaner UI, not only visually, but in code, too. A *lot* of
code was stripped from options/highlighting.js because it isn't needed anymore.

The textbox is now multiline, so large groups are more manageable. Newlines are
also a separator, in addition to commas.

The up/down arrow icons use stock GTK icons on Linux, and the included
arrows.png file on other platforms.