Remove "at 3AM" from nightly info text
Funny story: beyondboredom.net's timezone is now UTC, which isn't affected by
daylight saving, so the nightlies are actually built at 4AM for me. Not to
mention that it was always inaccurate for any timezone other than PST anyway.

Also: change "should" to "can" - it's not as strong of a word.