Fix version functions
Trying to make an object sometimes behave like a string doesn't work very well.

When gamefox_lib.version is assigned to the |lastversion| variable, it
references the version object instead of copying the contents of the version
string. As a result the initial version checks (especially on a fresh install)
were broken. They're fixed now.

Also, in the version functions, check if the argument is actually undefined or
just an empty string. I'm not sure if this is a regression, but it was making
the functions use the current version (instead of the passed version) when they
shouldn't be.