Use getter/setter for gamefox_lib.version
This avoids having to set the variable at the end of lib.js (due to
gamefox_lib.prefs not being available while gamefox_lib is being defined). It
also keeps the version variable and the pref in sync, removing the need to set
the pref from overlay.js's startup code.