Remove flicker when moving groups
The flicker was caused by selecting the new position of the group before the
pref was written. The group that used to be at the position is loaded, then the
pref is updated and it loads the correct group. Those two rapid loads are
visible.

Fixed the problem by removing the selection of the listbox, updating the pref,
then selecting the new group position. That way no other group is loaded during
the process.