Smarter method of finding post date node
We don't always know what node contains the post date, since other scripts can
change the header. By looping through the nodes and searching for one that
matches the post date, we're less likely to mess up here.

The regex check is used to make sure it really is the post date, and not just a
node that also happens to contain "Posted".