Fix admin/mod/vip posts
I've made the postDateNode test regex more lenient. It covers the important
parts of the post date and it's very unlikely that anything else we encounter
will match it.

Another unrelated problem that's fixed with this commit occurs only when the TC
is an admin/mod/vip. When this is the case, the status passed to the
highlighting functions is actually an array. I forgot about this case and made
searchStatus() blindly try to trim() the status, which doesn't work with an
array.