Add GameFAQs clock to QuickPost
I couldn't decide what to do with the clock in the header, so I just left it.
There's now two clock elements - one in the header and one for QuickPost (on
the topic and message list).

span#gamefox-clock becomes span.gamefox-clock (the CSS doc was actually wrong,
saying that we had span.gamefox-clock the whole time).

I made the floating QuickPost slightly wider to fit the clock.