Add "Themes" category to stylesheets
To make themes easier to notice, I've moved them to their own category at the
top of the style options. The "GameFAQs" category has been renamed to "Other",
and the "GameFOX" category is now at the bottom of the list (though above user
imports), since it's not that important to disable/enable stylesheets in it.