Improve position of topic QuickPost
Topic QuickPost is now perfectly centered horizontally, and almost centered
vertically. It's offset down a bit to make room for the character map.

Note: using window.innerWidth and window.innerHeight isn't correct here.
Because we are a Firefox extension, window.inner(Width|Height) holds the
dimensions of the entire window, including chrome. This isn't useful when
trying to center something in a web page. For that we have to use
doc.defaultView.inner(Width|Height) to get the actual viewport dimensions.