Detect browser startup vs new window
Count the number of open windows to make sure our startup code only runs once.
On startup, the number of open windows should be one.

Without this patch, all of gamefox_loader() is run every time a new window is
opened (kind of annoying to update tracked topics every time this happens).

I tested this when session restore is restoring multiple windows, and it still
works.