Import msgsPerPage on first run
I still want to keep this behavior, since it guarantees (as long as the user is
logged in) that we'll have the correct msgsPerPage setting as early as
possible.

In the future, I may even increase the frequency that GameFOX updates this
setting because it's the most reliable method we've got. I don't know how often
people will load a multi-page topic not on the last page, and if they'll do it
before they notice any glitches due to an incorrect msgsPerPage setting.