Better checks for disabling highlighting menulists
Since the extra type checkboxes for highlighting can apply to both topics and
messages, it's not enough to only check the group type to determine which
menulists to disable.

For example, if a group is highlighting topics that contain something in their
title, the message actions menulist would be disabled. If the TC checkbox is
checked, that menulist should be re-enabled because messages from the TC will
be matched by the group.