Fix date format + pagination prefwindow ids
The date format options' prefwindow and prefpane had the same id, which was
causing an exception to be logged in the error console every time it opened.
Unfortunately, it looks like the id is stored in localstore.rdf, so that
exception persisted until I manually deleted the lines from localstore.rdf. Not
much I can do about that, but at least the GameFOX bug is fixed.

Also, I updated the ids of the pagination window, which still referred to
Firefox's JS prefwindow which I copied the code from. Can't believe that was
still there. >_>