Fix tracking note dismiss link + wordwrap (again)
In some cases increasing the line height (from .8em to the height of the box)
will trigger wordwrap. To work around this, I've set the default line height in
the CSS to an unreasonably high number so GameFOX will always be decreasing it,
which should never cause wordwrap.

The number itself doesn't matter because it's fixed immediately after the link
is appended to the DOM.