Remove msgsPerPage auto-updating feature
This didn't work as well as intended, so I'm removing it.

I added the setting back to the options window, but as a small entry on the
Main -> Page pane, and without any of the other board settings. This is the
old way and it worked well enough at least.

There is no sync button anymore. I think basically everyone knows how many
messages they have per page, and it's not worth the UI bulk of having another
button added for one setting (as opposed to all of them with the previous board
settings pane). Of course, I could be wrong on this, so it's easy to just use
the function that already syncs it on first run.