Only disable <style> elements in the body
Takato emailed me about GameFOX disabling a <style> element added by his
script. Since the element is added to the head by his script, and as far as I
know all of the ad style elements are in the body, I'm attempting to avoid this
by only looking in document.body for the elements.