Move isLoggedIn() check to sig import function
By doing this, we can display the import error message in the signature pane on
first run instead of the current behavior, which is to leave the sig box blank
with no notification.

The extra login check remains in case gamefox_lib.isLoggedIn() misses
something, or as a catch-all error if the HTML of the page changes.