Beta compatibility
For our content policy, the host is dynamically set depending on the host of
the request origin. This keeps ad blocking and CSS disabling working on both
the main site and the beta.

The gamefox.useBeta pref is set automatically in gamefox_page.process().
gamefox_lib.domain is now a getter that returns based on this pref. This is
very useful because it means that simply using the beta will switch everything
over to opening beta links (QuickPosting, opening tracked topics/favorite
boards, etc), and going back to the main site will switch everything back.

A lot of XPath expressions were changed to start with // due to the addition of
some extra divs in the tree.

GameFAQs no longer outputs a message when tracking/untracking, so we have to
check if the destination of the tracking link has changed instead.