Support for new beta ad settings
These settings basically restore the board links to the way they were before
the beta, so simply checking if the page is on beta.gamefaqs.com won't work
anymore. Instead, we check whether userNav or userPanel exists. This should
also be forward-compatible without any changes for when the main site is
updated, too.