Add sig box to post.php, re-enable other features
The usual features weren't being added to post.php because GameFOX couldn't
find the "message" textarea. Easily fixed by changing that to "messagetext".

The sig box is now added to this page, too. One problem with it that's not
easily fixed is that if someone previews a post, then previews it again, the
custom sig will be lost (for example if the user made any changes to it before
previewing, or if another random sig was selected). This is because GameFAQs
doesn't save the custom sig anywhere in the form. We might be able to work
around this by detecting the sig in the message preview area.