Throttle unsuccessful attempts to fetch key
To prevent GameFOX from spamming GameFAQs with hits if they change the post
page so we can no longer find the key, disable key fetching for 10 hours if
more than 10 failed requests occur within an hour.

This won't affect people with lots of alts because, as long as everything is
working correctly, those requests will be successful and won't trigger
throttling.