Some post.php and ad CSS fixes
sky_col_wrap is gone, the multi-column divs are used for both legitimate and ad
reasons now, so we can't get rid of them. This is a problem on the post page,
because while the ad is hiddne, its container div isn't (and can't easily be).
We can still make the post page look pretty good by maximizing the space we can
use, though.