Add options button to toolbar automatically
Since the new menu in Firefox 4+ replaces the standard menu, our Tools ->
GameFOX location isn't as useful anymore. Adding a button to the toolbar
automatically seems to be the best solution to this, as several other
extensions do the same thing.

Also: change the options button icon to just be a standard logo. It looks
cleaner without the wrench icon in the corner, which was kind of pointless
anyway.