Fix bug when username is a JS object property
Usernames like "unwatch" will trigger this because "unwatch" is a property
inherited by all objects, including the object used to check if a username is
on a highlight list. By using hasOwnProperty(), we make sure that the property
is defined by us, and not just an inherited property.

hasOwnProperty() itself is safe from being overridden by us (from being in a
highlight list) because we convert all usernames to lowercase first
(-> "hasownproperty").