Fix login error with account manager
GameFAQs added a new hidden key field to the login form. We'll scrape the key
from the login error HTML source, store the key in the prefs, then try once
more to log in.

Also detect if the "CAPTCHA required" error is given, since otherwise you get
an incorrect password message from GameFOX, which is misleading.