Switch to new release notes system
I'm putting the last 10 (or so) version change logs on the same page.

This makes it better for people who skip versions when they upgrade (which is
pretty common when there's a rapid fire bugfix release). All the relevant
changes for multiple versions are in the same place and organized, which gets
rid of the weird merging of multiple versions into a single release notes page
during those rapidfire releases.

Also, this makes it much easier to release new versions on a timely basis,
because the overhead for doing so is now lower - just a simple copy/paste of
the latest NEWS into this file, without the need to create a new file and write
a long description of the update. That usually ended up being an awkward
rephrasing of the NEWS, anyway.

The ?version= parameter is used by the server to check if the version is out of
date. Firefox auto-downloads XPI as soon as they're available from the update
RDF now, so it's possible to upgrade to an out-of-date version and not get the
changelog shown until Firefox restarts later, after a new version has already
been released. This catches those cases.