Add changes-next.mdwn file for next version
Since everything is in a single file now, we couldn't upload the changes for
the next version without interfering with the current version's page. To fix
this, I've added changes-next.mdwn as a staging area for the next version.

It can be previewed at http://beyondboredom.net/gamefox/changes-next

When running "make upload", changes-next.mdwn is automatically copied to
changes.mdwn, so people who are updating will always be shown the correct
information.