Fix use of double slashes in XPath expressions
When using a context node other than the document, an expression that starts
with "//" will search the entire DOM, not just the elements below the context.
To fix this, we should use ".//" instead.

As an example of what this fixed, if a link in a topic contains "track", and
the topic is archived (so the actual GameFAQs tracking link doesn't appear),
then that link will get the event listener added to it. This means the link
can't be opened and an alert box pops up about tracking when it's clicked.

By changing the XPath to use ".//", this makes sure the only tracking link we
find is inside the user panel/nav, which is what we intended to do in the first
place by setting a context node.