Fix sidebar login form, accounts.js refactoring
I figured the easiest way to solve this is to make the login form use the same
code as the account manager, so I decoupled the actual login logic from the
account manager stuff. A callback function is used to handle the results for
either the sidebar or the account manager.

Middle clicking the login button to open a new tab no longer works. I don't
really think this feature is worth the trouble to keep, although if someone
complains I'll try to add it back.