Fix quoting when Disconnect extension is installed
Disconnect adds a <tr> and <td> to the DOM, so GameFOX thinks that the page is
using left message data, even if it's not. This causes quoting to break with
above message data.

Making sure getMsgDataDisplay() only looks in div#content fixes this, since
Disconnect isn't adding elements in there (and has no reason to).