Fix QuickWhois hide-on-hover bug
We need to always set the mouseout event listener here - the code will handle
any irrelevant events as they occur. Failing to set the event here will mean
that if the QuickWhois box is initially created via double click, toggling it
later with hover will cause it to permanently stay open because there's nothing
listening for the mouseout event.