Don't import the sig on first run
The default is now going to be to use the GameFAQs signature (i.e., no
custom_sig value). This is more intuitive to people who are used to managing
their signatures from GameFAQs, and people who want to use GameFOX instead can
click the "Grab from GameFAQs" button or type in a custom sig to have the same
functionality.

Also: Remove the "Update GameFOX Signature" button on the sigquote page, since
this isn't going to be necessary anymore when GameFAQs is managing sigs.