Fix incorrect character count with quick edit
I noticed that the initial character count was always incorrect, since typing a
single character in the box would actually make the count decrease. Turns out
this is because the message has \r\n newlines, and Firefox converts them into \n
newlines automatically - but not before the first count is done.

Converting the newlines ourselves fixes this bug.