Fix double clicking <td> to open QuickWhois
QuickWhois was having a problem finding the message header, since it used
findParent. This ignores the current element (skips straight to parentNode), but
in this case, the current element is what we're searching for in the first
place.

Some parts of the code add an additional check to see if the current element is
what they want. I've decided to simplify this by incorporating that into the
function, and renaming it to findClosest (similar to jQuery's closest()).