Don't open the options window on first run
Since we no longer have any first run code in the option window, we don't need
to open it anymore. This eliminates the "in your face" experience of installing
GameFOX - a window opening that the user is forced to dismiss is a bit annoying.

For discoverability, we have the toolbar icon auto-added already, so it's easy
to find the options window.