Fix QuickWhois glitch when triggered from <td>
The old method of searching for a tag name isn't reliable enough, since there
are multiple <td>s due to QuickWhois itself, and searching for <div> will fail
if it starts at the post header node (which is a <td>).

To fix this, I've extended findClosest to support a test function, which can
make arbitrary tests to see if the node is correct. That lets us find the
correct node with much more accuracy.