Add friend + PM buttons to QuickWhois
These are part of the standard profile, so we could just add them to to the list
of fields to fetch. But the buttons look a bit ugly in the otherwise compact
QuickWhois. So, I'm doing special formatting for them and displaying them as
tiny Silk icons next to the username.