Fix quoting after opening QuickWhois in some cases
When QuickWhois is opened in the first post with above msg data, the
getMsgDataDisplay() function misinterprets it as being left msg, which breaks
quoting. I updated the function to be more reliable in this case (by only
checking immediate children rather than all <td>s).

I got rid of the try/catch block because if getMsgDataDisplay() throws an
exception, something is probably really wrong, and we shouldn't try to hide it.