Fix getAccountName() with clock
Previously, the first text node of the login box was used to get the username.
The clock wasn't included in this since it was a separate element. However, we
want to be flexible enough to allow foreign elements (such as <a>), so I think a
good compromise is to cut off all text after the "|" (clock separator). We
control the separator (so it won't change from under us), so I think it's an
acceptable solution.