Switch account manager from cookies to passwords
Yep, it's finally been implemented. The account manager will now store passwords
of all new and updated accounts. Existing cookie accounts will continue to work
- when the cookies expire, and the password is re-entered, it'll just be
converted into a password-storing account, until eventually the passwords for
all accounts are stored. So, it's a gradual, seamless upgrade process.

Firefox's Login Manager is used for this, so we should be good in terms of
security.

You do lose a little bit of speed this way (there are extra requests to GameFAQs
which you don't need to do with the cookie swapping method), but I think it's
worth it to not need to re-enter all your accounts' passwords every 30 days or
so.