Bail out of QP keyObserver is gf_lib is undefined
To be honest, I'm not really sure what's causing this, but I need to do
something about the "gamefox_lib is undefined" message being spammed in the
error console.

As far as I can tell the function is still working properly in most cases. And
if gamefox_lib isn't defined then there isn't a whole lot we can do in that kind
of environment, anyway. So I'm going to go from failing loudly here to failing
silently.