Fix QuickPost with "ad at bottom"
In multi-page topics with this ad setting, QuickPost would appear above
the page jumper/board nav, which is weird. Moving it out to be a direct
child of #content fixes this, but the padding is different so it's no
longer aligned with the message list - adding the col_layout class is
sort of hackish but seems like a robust fix.