Remove trimGFCodeNewlines function
This function wasn't working anyway since the regex hasn't been updated
to the new quote tags, but I'm removing it instead of updating it. Quote
tags make this a GameFAQs issue rather than a GameFOX one, since all
users will see the same amount of whitespace now, with or without
GameFOX. (This wasn't the case before due to the p tags.) For the sake
of simplicity I don't feel it worth it to enforce this anymore.