Fix topic creator label
User statuses are now converted to array. This isn't needed in most
cases, but for topics where the TC is also an admin, mod, or VIP, it
makes things a lot easier.

There's a little bit of kludgery to get the "(Topic Creator)" text
removed so we can keep our "(tc)" label the same (and in the same
place) in the form of the showStatus() function - again, it's not needed
in most cases, but in places where an admin/mod/VIP is also a TC, it's
easier to just completely rebuild the user status text so things look
right.