Remove TC from various functions
TC is now combined with the status argument for highlighting, so none of
the highlighting functions need a separate variable for it anymore. (For
simplicity, support for "(tc)" as a special-cased username to highlight
the TC has been removed, but that was deprecated years ago, so it
shouldn't affect anyone.)

Removed it from newURI() as well - it was already removed from the
actual URI it generated in a previous commit, but the function argument
lingered. Now that's removed, which required updating all function calls
to newURI().