Add signature spans when using edit history viewer
Move sig span adding into its own function. I don't want to pollute
utils.js with esoteric functions like this, so I made a new page_scripts
file for it (which is where it would probably be along with all the
other message list processing if that ever happens).

This fixes these edit history viewer bugs: hide/remove signatures CSS
not working, quoting not removing signatures (since it relies on the sig
span instead of regex now).

I also removed the sigspans toggle pref now that quoting relies on it. I
don't think anyone is using the pref, and it's better to avoid bug
reports about quoting not removing sigs if someone is using it.

One quoting quirk remains: the signature is not filtered out in
selection quoting, since the HTML of the selection isn't available. This
could be considered a feature, though, since the user might want to
actually include the signature in the quote if they explicitly select
it. It's probably okay to leave it.