Fix char map/HTML buttons pref check
The pref names for extended/gfcode buttons were changed/removed but
weren't updated in the pref check function, so it threw an exception if
it reached them (since it's one big OR statement, the offending prefs
were skipped if basic HTML buttons or character map were left enabled).

Also, give the function a better name and change it to a getter.