Fix adding/removing accounts from sidebar
The name of the string bundle in the sidebar still uses the generic
overlay-strings id, which won't cause any namespace conflicts since the
sidebar is its own isolated thing, but the account prompts are expecting
gamefox-overlay-strings because that's what it's named in the browser
overlay. Rename the sidebar one to match the browser overlay one for
consistency.