Fix tracking JSON error breaking quick post
JSON.parse throws an error if the string is blank, so set the default
for the tracking list JSON pref to be "{}" like the other JSON prefs.
This is a simple fix, but it isn't very robust if this pref (or any
other JSON pref) gets set to something invalid - another fix that
handles these cases is probably needed.