Replace third party cookie workarounds with fix
The result of bug 437174 was a forceAllowThirdPartyCookie flag on HTTP
requests. This was added in Gecko 1.9.2 (Firefox 3.6). Using this
simplifies our support for having third party cookies disabled - we only
need to set one attribute and we're good, and there's no need to have
the third party cookie warning prompt anymore, either.