Match nonbreaking space in post date regex
GameFAQs now uses a nonbreaking space between the time and the AM/PM
label, which breaks our regex for finding the post date node. This is a
silly bug, and could potentially happen again somewhere else, but I'm
only changing the one space character in this particular spot to "\s"
for now. Thanks to people on Blood Money for discovering this problem.