Show a notification when an error is thrown
Display the exception thrown when GameFOX encounters an error at the top
of the tab, using the browser's notificationbox. This info is shown in
the Error Console, but most users don't include that with bug reports,
and the actual error is usually hidden among other harmless errors.
Showing the error directly to the user should hopefully make it easier
to fix issues.

To prevent this from being too annoying, there's an "Ignore" button that
will hide all error notifications until the user upgrades (or
downgrades) GameFOX.