Fix finding post buttons on post.php
form.elements.namedItem()'s behavior when there is more than one element
with the same name is sort of odd. MDN says that Firefox 8+ will return
the first element with the name, but testing on Fx14 shows that it
returns an HTMLCollection of all the matching elements. (But another
user with Fx14 has the opposite experience.) Conversely, Fx3.6 only
returns a single element for me. To avoid this mess, use
doc.getElementsByName(). (I'd use form.getElementsByName(), but
JavaScript only defines that function for the document for some reason.)