Fix edit link appearing when it shouldn't
This is the nbsp issue again - it was fixed only for custom date
formats before, not any other part that uses the postDate variable.
Replace the nbsp earlier in the code, but ideally after postDateElement
is created to avoid clobbering the nbsp on the page for people who don't
use custom date formats.